% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateModel.R
\name{CreateModel}
\alias{CreateModel}
\title{Creates an SVM model from training set}
\usage{
CreateModel(TrainingSet)
}
\arguments{
\item{TrainingSet}{List input that has training data set and prediction data set.}
}
\value{
The SVM model
}
\description{
The function takes in the training set to create a model that clusters good worms and bad worms together.
}
