% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPred.R
\name{PlotPred}
\alias{PlotPred}
\title{Plots the data set, the model and the prediction results}
\usage{
PlotPred(Prediction, StageData)
}
\arguments{
\item{Prediction}{Prediction results from the RunModel function}

\item{StageData}{Stage Data containing the data set of worms in one stage (e.g. Adult)}
}
\value{
Returns three plots
}
\description{
The function takes in the prediction results and the Stage data and plots them. The resulting plots are as follows:
1- Plots all the worm data 2- Plots the predicted bad worms 3- Plots the predicted good worms
}
