% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunModel.R
\name{RunModel}
\alias{RunModel}
\title{Runs the model againts a data set}
\usage{
RunModel(StageData, Model)
}
\arguments{
\item{StageData}{Stage Data containing the data set of worms in one stage (e.g. Adult)}

\item{Model}{The SVM model}
}
\value{
Returns the prediction of this model
}
\description{
The function takes in an SVM model and a worm data set and outputs a categorized list of worms.
}
