% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WormCount.R
\name{WormCount}
\alias{WormCount}
\title{Counts number of worms}
\usage{
WormCount(StageData, WormIDs)
}
\arguments{
\item{StageData}{Stage Data containing the data set of worms in one stage (e.g. Adult)}

\item{WormIDs}{File containing annotated worm IDs of the channel data.}
}
\value{
A string containing the number of good and bad worms
}
\description{
The function takes in the channel data and the wormIDs and counts how many good worms and bad worms in the specfied
stage are in the provided annotated set.
}
